% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance-dist.R
\name{abundance_dist}
\alias{abundance_dist}
\title{Distribution of abundance along the environmental gradient}
\usage{
abundance_dist(
  fit,
  seed = NULL,
  ndraws = 50,
  summaries = c("mean", "median", "q90"),
  include_zero_inflation = TRUE
)
}
\arguments{
\item{fit}{modskurt_fit object}

\item{seed}{optional seed for ndraws sample}

\item{ndraws}{optional subset of posterior draws}

\item{summaries}{character vector of distribution summarise, see details}

\item{include_zero_inflation}{where to include zero-inflation in summaries}
}
\value{
a ggplot object
}
\description{
Distribution of abundance along the environmental gradient
}
