% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance-plots.R
\name{abundance_range}
\alias{abundance_range}
\title{Assess abundance range along x using distributional measures}
\usage{
abundance_range(
  fit,
  seed = NULL,
  ndraws = 50,
  capture_pct = 80,
  using_range = "HAZ",
  based_on = "median",
  region = "centre",
  include_zero_inflation = FALSE,
  plotted = FALSE,
  point_alpha = 0.2,
  line_alpha = 0.02,
  range_colour = "red"
)
}
\arguments{
\item{fit}{modskurt_fit object}

\item{seed}{optional seed for ndraws sample}

\item{ndraws}{optional subset of posterior draws}

\item{capture_pct}{percent of total abundance density (most_dens) or percent
tolerance from highest abundance summary (high_zone), based on "based_on"
summary of abundance}

\item{using_range}{"ADL" for highest abundance density limit, or
"HAZ" for highest abundance zone, see details}

\item{based_on}{which distribution summary to use for density or high zone,
one of "mean", "median", or any quantile proportion as "q" + pecentage
(e.g. "q90" = 90\% quantile)}

\item{region}{one-sided threshold ("left", "right") or central range
("centre")}

\item{include_zero_inflation}{whether to use expectations and quantiles from
nb or zinbl}

\item{plotted}{plot on the abundance distribution}

\item{point_alpha}{opacity of data points, default 0.1 is good for ~ 1000 obs}

\item{line_alpha}{opacity of summary lines, default 0.1 is good for ~ 50
draws}

\item{range_colour}{colour of range geometries}
}
\value{
a matrix of ranges, or a ggplot object if plotted is TRUE
}
\description{
Assess abundance range along x using distributional measures
}
