% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroinvertebrates.R
\docType{data}
\name{aonides}
\alias{aonides}
\title{*Aonides trifida* estuarine abundance}
\format{
## `aonides`
A data frame with 251 rows and 7 columns:
\describe{
  \item{estu}{Estuary where the count was observed}
  \item{year}{Year of survey}
  \item{count}{Total counted in a 250 gram sediment core}
  \item{mud_pct}{\% of sediment that was muddy (<63microns?) at the time of sampling}
  \item{tn_conc}{total nitrogen concentration (mg/L) at the time of sampling}
  \item{toc_pct}{total organic carbon content (\%) at the time of sampling}
  \item{tp_conc}{total phosphorus concentration (mg/L) at the time of sampling}
}
}
\source{
<https://www.saltecology.co.nz/>
}
\usage{
data(aonides)
}
\description{
Discrete counts of aonides trifida in intertidal estuaries of New Zealand
}
\keyword{datasets}
