% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-computation.R
\name{check_computation}
\alias{check_computation}
\title{Check posterior computation is acceptable}
\usage{
check_computation(fit)
}
\arguments{
\item{fit}{modskurt_fit object}
}
\value{
a list with ...
}
\description{
Check posterior computation is acceptable
}
