% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-posterior.R
\name{check_post_calibration}
\alias{check_post_calibration}
\title{Check prediction calibration using discrete PIT histograms}
\usage{
check_post_calibration(fit, ndraws = 50)
}
\arguments{
\item{fit}{fitted model object from `mskt_fit`}

\item{ndraws}{number of draws from the joint posterior to use}
}
\description{
Check prediction calibration using discrete PIT histograms
}
