% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-posterior.R
\name{check_post_dens}
\alias{check_post_dens}
\title{visual display of chain mixing and prior-data conflicts}
\usage{
check_post_dens(
  fit,
  pars = attr(spec, "pars"),
  prob_inner = 0.5,
  prob_outer = 0.99,
  by_chain = FALSE,
  chain_legend_ncol = 1,
  plot = TRUE
)
}
\arguments{
\item{fit}{fitted model object from `mskt_fit`}

\item{pars}{parameters to plot densities for, default is all}

\item{prob_inner}{inner probability (high density interval) to highlight}

\item{prob_outer}{outer probability range to plot the density between}

\item{by_chain}{plot chains separately to identify multiple modes and mixing
issues}

\item{chain_legend_ncol}{for the thesis figure}

\item{plot}{boolean, plot together, or return list of individual ggplots}
}
\description{
visual display of chain mixing and prior-data conflicts
}
