% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-posterior.R
\name{check_post_influencers}
\alias{check_post_influencers}
\title{Check for posterior misspecification by assessing influence data points}
\usage{
check_post_influencers(fit)
}
\arguments{
\item{fit}{fitted model object from `mskt_fit`}
}
\description{
Check for posterior misspecification by assessing influence data points
}
