% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-prior.R
\name{check_prior_dens}
\alias{check_prior_dens}
\title{Check marginal prior densities are encoded correctly}
\usage{
check_prior_dens(
  spec,
  pars = attr(spec, "pars"),
  use_subset = TRUE,
  prob_inner = 0.5,
  prob_outer = 0.99
)
}
\arguments{
\item{spec}{the modskurt model specification returned by `mskt_spec`}

\item{pars}{parameters to plot densities for, default is all}

\item{use_subset}{whether to use specified subset or all data}

\item{prob_inner}{inner probability (high density interval) to highlight}

\item{prob_outer}{outer probability range to plot the density between}
}
\value{
a patchwork of ggplots
}
\description{
Check marginal prior densities are encoded correctly
}
