% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-prior.R
\name{check_prior_dist}
\alias{check_prior_dist}
\title{Check prior distribution statistics for potential misspecification clues}
\usage{
check_prior_dist(spec, use_subset = TRUE, n_draws = 50)
}
\arguments{
\item{spec}{the modskurt model specification returned by `mskt_spec`}

\item{use_subset}{whether to use specified subset or all data}

\item{n_draws}{number of draws from the joint prior to use}
}
\description{
Check prior distribution statistics for potential misspecification clues
}
