% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-prior.R
\name{check_prior_mskt}
\alias{check_prior_mskt}
\title{Check prior predictions of the modskurt mean cover all possible shapes}
\usage{
check_prior_mskt(spec, use_subset = TRUE, n_draws = 50)
}
\arguments{
\item{spec}{the modskurt model specification returned by `mskt_spec`}

\item{use_subset}{whether to use specified subset or all data}

\item{n_draws}{number of draws from the joint prior to use}
}
\value{
a ggplot object
}
\description{
Check prior predictions of the modskurt mean cover all possible shapes
}
