% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-prior.R
\name{check_prior_zero}
\alias{check_prior_zero}
\title{Check prior predictions of excess zero probabilities cover all possible relationships}
\usage{
check_prior_zero(spec, use_subset = TRUE, n_draws = 50)
}
\arguments{
\item{spec}{the modskurt model specification returned by `mskt_spec`}

\item{use_subset}{whether to use specified subset or all data}

\item{n_draws}{number of draws from the joint prior to use

Plot prob zero for median mu_n? or just for a bell curve to show how it works?}
}
\description{
Check prior predictions of excess zero probabilities cover all possible relationships
}
