% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-stanmodels.R
\name{compile_stanmodels}
\alias{compile_stanmodels}
\title{Compile ModSkurt Stan Models}
\usage{
compile_stanmodels(force_recompile = FALSE)
}
\arguments{
\item{force_recompile}{whether to recompile even if nothing has changed}
}
\description{
Compile ModSkurt Stan Models
}
