% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative-binomial.R
\name{dzinb2}
\alias{dzinb2}
\title{Zero-inflated Negative Binomial Probability Mass}
\usage{
dzinb2(x, mu, kappa, zi, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) values}

\item{mu}{mean of the distribution}

\item{kappa}{dispersion (inverse square root size)}

\item{zi}{vector of (real lying in \[0, 1\]) zero-inflation parameters}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p)}
}
\value{
The (log) probability mass at x, given mu, kappa and zi
}
\description{
Zero-inflated Negative Binomial Probability Mass
}
\keyword{internal}
