% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative-binomial.R
\name{kappa_to_phi}
\alias{kappa_to_phi}
\title{Convert overdispersion kappa to reciprocal dispersion phi}
\usage{
kappa_to_phi(kappa)
}
\arguments{
\item{kappa}{overdispersion parameter}
}
\description{
Convert overdispersion kappa to reciprocal dispersion phi
}
