% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modskurt-fit.R
\name{modskurt_fit}
\alias{modskurt_fit}
\title{Fit the specified ModSkurt model by sampling its posterior distribution}
\usage{
modskurt_fit(
  spec,
  seed = 1234,
  init = NULL,
  chains = 6L,
  parallel_chains = 6L,
  iter_warmup = NULL,
  iter_sampling = NULL,
  max_treedepth = NULL,
  adapt_delta = NULL,
  show_messages = TRUE,
  show_exceptions = FALSE,
  ...
)
}
\arguments{
\item{spec}{the ModSkurt model specification returned by `modskurt_spec`}

\item{seed}{a seed for the Stan random number generator}

\item{init}{initialisation method, see ?cmdstanr::`model-method-sample`}

\item{chains}{the number of Markov chains}

\item{parallel_chains}{how many cores to distribute chains over}

\item{iter_warmup}{number of iterations used to learn posterior curvature}

\item{iter_sampling}{number of posterior samples drawn}

\item{max_treedepth, adapt_delta}{tuning parameters, see details}

\item{show_messages}{additional verbosity}

\item{show_exceptions}{additional verbosity}

\item{...}{additional arguments to ?cmdstanr::`model-method-sample`}
}
\value{
A `cmdstanr::CmdStanMCMC` object
}
\description{
Fit the specified ModSkurt model by sampling its posterior distribution
}
