% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskt-sim.R
\name{mskt}
\alias{mskt}
\title{Log-Modskurt (mskt) unimodal mean function}
\usage{
mskt(x, H, m, s, r = 0.5, d = 0, p = 1)
}
\arguments{
\item{x}{predictor variable (real number vector)}

\item{H}{height of curve at mode (real number > 0)}

\item{m}{mode (real number)}

\item{s}{scale (real number > 0)}

\item{r}{asymmetry (real number in (0, 1))}

\item{d}{flatness (real number >= 0)}

\item{p}{tail exaggeration (real number in (0, 2))}
}
\value{
vector of log-scale mean values (real number vector)
}
\description{
Log-Modskurt (mskt) unimodal mean function
}
\keyword{internal}
