% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskt-fit.R
\name{mskt_fit}
\alias{mskt_fit}
\title{Fit the specified modskurt model by sampling its posterior distribution}
\usage{
mskt_fit(
  spec,
  train_prop = 1,
  train_seed = NULL,
  seed = NULL,
  init = NULL,
  chains = 4,
  parallel_chains = getOption("mc.cores", 1),
  iter_warmup = 1000,
  iter_sampling = 1000,
  max_treedepth = NULL,
  adapt_delta = NULL,
  show_messages = TRUE,
  show_exceptions = FALSE,
  ...
)
}
\arguments{
\item{spec}{the modskurt model specification returned by `mskt_spec`}

\item{train_prop}{the proportion of data to train the posterior on}

\item{train_seed}{a seed for the training sample of proportion `train_prop`}

\item{seed}{a seed for the Stan random number generator}

\item{init}{initialisation method, see ?cmdstanr::`model-method-sample`}

\item{chains}{the number of Markov chains}

\item{parallel_chains}{how many cores to distribute chains over}

\item{iter_warmup}{number of iterations used to learn posterior curvature}

\item{iter_sampling}{number of posterior samples drawn}

\item{max_treedepth, adapt_delta}{tuning parameters, see details}

\item{show_messages}{additional verbosity}

\item{show_exceptions}{additional verbosity}

\item{...}{additional arguments to ?cmdstanr::`model-method-sample`}
}
\value{
A `cmdstanr::CmdStanMCMC` object
}
\description{
Fit the specified modskurt model by sampling its posterior distribution
}
