% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskt-fit.R
\name{mskt_inits}
\alias{mskt_inits}
\title{internal only for initialising from prior}
\usage{
mskt_inits(sdat, chains, from_data = TRUE)
}
\description{
internal only for initialising from prior
}
\keyword{internal}
