% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskt-sim.R
\name{mskt_sim_nb}
\alias{mskt_sim_nb}
\title{Simulate negative binomial data with a modskurt mean curve}
\usage{
mskt_sim_nb(
  x,
  n_per_x = 1,
  H,
  m,
  s,
  r = 0.5,
  d = 0,
  p = 1,
  kap,
  g0 = -Inf,
  g1 = 0
)
}
\arguments{
\item{x}{vector of environmental gradient values}

\item{n_per_x}{number of y replicates to simulate for each x value}

\item{H}{height}

\item{m}{middle}

\item{s}{spread}

\item{r}{assymetry}

\item{d}{flatness}

\item{p}{persistence}

\item{kap}{negative binomial dispersion}

\item{g0}{constant rate of zero-inflation}

\item{g1}{decay of zero-inflation as mean increases}
}
\value{
data.frame with simulated mu, zi, and y for each x
}
\description{
Simulate negative binomial data with a modskurt mean curve
}
