% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskt-spec.R
\name{mskt_spec}
\alias{mskt_spec}
\title{Specify ModSkurt model for the distribution of `y` over `x`}
\usage{
mskt_spec(
  data,
  y,
  x,
  effort = NULL,
  dist = "nb",
  shape = "rdp",
  hyperparams = list(),
  pred_grid = NULL
)
}
\arguments{
\item{data}{data.frame or tibble}

\item{y}{species count, optionally named - see details}

\item{x}{environmental gradient, optionally named - see details}

\item{effort}{sampling effort used to observe y}

\item{dist}{distribution of y over x}

\item{shape}{shape of distribution mean - see ?mskt_shape()}

\item{hyperparams}{list of hyper-parameters to adjust in-built priors}

\item{pred_grid}{vector of x values to make y predictions over}
}
\value{
list of standata for prior verification and posterior estimation
}
\description{
Specify ModSkurt model for the distribution of `y` over `x`
}
\details{
## Forms of the discrete (count) data distribution
- `"nb"`: Negative Binomial
- `"zinbl"`: Zero-inflated Negative Binomial (Linked)

An interactive graph of these count distributions and possible parameters is
available at [https://salt-ecology.shinyapps.io/nb-zinbl-prior-specs/]().

## Shapes for the mean function
- default: height (H), mode (m), scale (s)
- optional extras
  + `"r"` assymmetry or skew
  + `"d"` flatness or peakedness
  + `"p"` tail exaggeration or pinching

These can be combined as desired, e.g. `"rp"`, `"rdp"`, ...

An interactive graph of the modskurt mean function and possible parameters is
available at [https://salt-ecology.shinyapps.io/modskurt-prior-specs/]().
}
