% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative-binomial.R
\name{pzinb2}
\alias{pzinb2}
\title{Zero-inflated Negative Binomial Distribution Function}
\usage{
pzinb2(q, mu, kap, zi)
}
\arguments{
\item{q}{vector of quantiles}

\item{mu}{mean of the distribution}

\item{zi}{vector of (in \eqn{[0, 1]}) zero-inflation parameters}

\item{kappa}{dispersion (inverse square root size)}
}
\description{
Zero-inflated Negative Binomial Distribution Function
}
\keyword{internal}
