% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative-binomial.R
\name{qnbinom2}
\alias{qnbinom2}
\title{The Negative Binomial Quantile Function (alternative parameterisation)}
\usage{
qnbinom2(p, mu, kappa, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{mu}{mean}

\item{kappa}{dispersion (inverse square root size)}

\item{lower.tail}{logical; if TRUE (default), probablities are \eqn{P[X <= x]}, otherwise, \eqn{[X > x]}}

\item{log.p}{logical; if TRUE, probabilities are given as log(p)}
}
\value{
A vector of quantiles, each of which correspond to a probability in p
}
\description{
The Negative Binomial Quantile Function (alternative parameterisation)
}
\keyword{internal}
