% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative-binomial.R
\name{qzinb2}
\alias{qzinb2}
\title{Zero-inflated Negative Binomial Quantile Function}
\usage{
qzinb2(p, mu, kappa, zi, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of quantile probabilities}

\item{mu}{mean of the distribution}

\item{kappa}{dispersion (inverse square root size)}

\item{zi}{vector of (in \eqn{[0, 1]}) zero-inflation parameters}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X <= x]}, otherwise, \eqn{P[X > x]}}

\item{log.p}{logical; if TRUE, probabilities p. This doesn't affect zi}
}
\value{
A vector of quantiles, each of which coincide with the respective probability in p
}
\description{
Zero-inflated Negative Binomial Quantile Function
}
\keyword{internal}
