% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redstripe.R
\docType{data}
\name{redstripe}
\alias{redstripe}
\title{*Sebastes proriger* (redstripe rockfish) estuarine abundance}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2330 rows and 13 columns.
}
\source{
<https://www.fisheries.noaa.gov/west-coast/science-data/us-west-coast-groundfish-bottom-trawl-survey>
}
\usage{
data(redstripe)
}
\description{
Discrete counts of *sebastes proriger* along the Pacific Northwest continental slope and shelf
}
\keyword{datasets}
