% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuangi.R
\docType{data}
\name{tuangi}
\alias{tuangi}
\title{Tuangi (*Austrovenus structburyi*, NZ Cockle) estuarine abundance}
\format{
## `tuangi`
A data frame with 764 rows and 11 columns:
\describe{
  \item{year}{Year of survey}
  \item{lon, lat}{Longitude and latitude of the sample}
  \item{count}{Total tuangi counted in a 250 gram sediment core}
  \item{mud_pct}{\% of sediment that was muddy (<63microns?) at the time of sampling}
  \item{tn_conc}{total nitrogen concentration (mg/L) at the time of sampling}
  \item{toc_pct}{total organic carbon content (\%) at the time of sampling}
  \item{tp_conc}{total phosphorus concentration (mg/L) at the time of sampling}
  \item{sst_min, sst_avg, sst_max}{Min, mean, and max sea surface temperature (deg C) at the location of sampling over the 12 months preceding}
}
}
\source{
<https://www.saltecology.co.nz/>
}
\usage{
data(tuangi)
}
\description{
Discrete counts of tuangi in intertidal estuaries of New Zealand
}
\keyword{datasets}
