% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskt-sim.R
\name{zilink}
\alias{zilink}
\title{ZINBL zero-inflation link function}
\usage{
zilink(mu, H, g0, g1)
}
\arguments{
\item{mu}{mean negative binomial abundance (real number vector)}

\item{H}{height of curve at mode (real number > 0)}

\item{g0}{log-odds probability of excess zero when mu = 0}

\item{g1}{rate at which log-odds probability of excess zero decreases as mu increases}
}
\value{
vector of excess zero probabilities (real number vector in [0, 1])
}
\description{
ZINBL zero-inflation link function
}
\keyword{internal}
