% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance-plots.R
\name{abundance_dist}
\alias{abundance_dist}
\title{Distribution of abundance along the environmental gradient}
\usage{
abundance_dist(
  fit,
  seed = NULL,
  ndraws = 50,
  summaries = c("mean", "median", "q90"),
  include_zero_inflation = TRUE,
  point_alpha = 0.2,
  line_alpha = 0.1
)
}
\arguments{
\item{fit}{modskurt_fit object}

\item{seed}{optional seed for ndraws sample}

\item{ndraws}{optional subset of posterior draws}

\item{summaries}{character vector of distribution summarise, see details}

\item{include_zero_inflation}{whether to include zero-inflation in summaries}

\item{point_alpha}{opacity of data points, default 0.1 is good for ~ 1000 obs}

\item{line_alpha}{opacity of summary lines, default 0.1 is good for ~ 50 draws}
}
\value{
a ggplot object
}
\description{
Distribution of abundance along the environmental gradient
}
