% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-computation.R
\name{check_computation}
\alias{check_computation}
\title{Check posterior computation is acceptable}
\usage{
check_computation(fit, show_info = TRUE)
}
\arguments{
\item{fit}{fitted model object from `mskt_fit`}

\item{show_info}{print information about model spec and posterior sample}
}
\value{
a list with ...
}
\description{
Check posterior computation is acceptable
}
