% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-prior.R
\name{check_prior_dens}
\alias{check_prior_dens}
\title{Check marginal prior densities are encoded correctly}
\usage{
check_prior_dens(
  spec,
  train_prop = 1,
  train_seed = NULL,
  prob_inner = 0.5,
  prob_outer = 0.99
)
}
\arguments{
\item{spec}{the modskurt model specification returned by `mskt_spec`}

\item{train_prop}{the proportion of data to train the posterior on}

\item{train_seed}{a seed for the training sample of proportion `train_prop`}

\item{prob_inner}{inner probability (high density interval) to highlight}

\item{prob_outer}{outer probability range to plot the density between}
}
\value{
a patchwork of ggplots
}
\description{
Check marginal prior densities are encoded correctly
}
