% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative-binomial.R
\name{dnbinom2}
\alias{dnbinom2}
\title{The Negative Binomial Probability Mass Function (alternative parameterisation)}
\usage{
dnbinom2(x, mu, kappa, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles}

\item{mu}{mean}

\item{kappa}{dispersion (inverse square root size)}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
the (log) probability mass at x, given mu and kappa
}
\description{
The Negative Binomial Probability Mass Function (alternative parameterisation)
}
\keyword{internal}
